2025-05-14-16.25/requirements_simple.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
bcrypt==4.3.0
Pillow==10.0.0
PyMuPDF==1.22.5
pdf2image==1.16.3
python-docx==0.8.11
python-pptx==0.6.21
pytesseract==0.3.10
gunicorn==21.2.0
imap-tools==1.0.0
